backend/requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
markdown
django-filter
django-cors-headers
psycopg2-binary
dj-database-url